Updated form styles and functionality

+++ wp-content/plugins/edd-software-licensing/samples/edd-sample-theme/updater/theme-updater-admin.php
@@ -57,7 +57,7 @@ class EDD_Theme_Updater_Admin {
 		// Strings passed in from the updater config
 		$this->strings = $strings;
 
-		add_action( 'admin_init', array( $this, 'updater' ) );
+		add_action( 'init', array( $this, 'updater' ) );
 		add_action( 'admin_init', array( $this, 'register_option' ) );
 		add_action( 'admin_init', array( $this, 'license_action' ) );
 		add_action( 'admin_menu', array( $this, 'license_menu' ) );
@@ -72,6 +72,9 @@ class EDD_Theme_Updater_Admin {
 	 * since 1.0.0
 	 */
 	function updater() {
+		if ( ! current_user_can( 'manage_options' ) ) {
+			return;
+		}
 
 		/* If there is no valid license key status, don't allow updates. */
 		if ( get_option( $this->theme_slug . '_license_key_status', false) != 'valid' ) {
@@ -231,10 +234,8 @@ class EDD_Theme_Updater_Admin {
 
 		// Make sure the response came back okay.
 		if ( is_wp_error( $response ) ) {
-			return false;
+			wp_die( $response->get_error_message(), __( 'Error' ) . $response->get_error_code() );
 		}
-
-		$response = json_decode( wp_remote_retrieve_body( $response ) );
 
 		return $response;
 	 }
@@ -255,13 +256,85 @@ class EDD_Theme_Updater_Admin {
 			'item_name'  => urlencode( $this->item_name )
 		);
 
-		$license_data = $this->get_api_response( $api_params );
+		$response = $this->get_api_response( $api_params );
+
+		// make sure the response came back okay
+		if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
+
+			if ( is_wp_error( $response ) ) {
+				$message = $response->get_error_message();
+			} else {
+				$message = __( 'An error occurred, please try again.' );
+			}
+
+		} else {
+
+			$license_data = json_decode( wp_remote_retrieve_body( $response ) );
+
+			if ( false === $license_data->success ) {
+
+				switch( $license_data->error ) {
+
+					case 'expired' :
+
+						$message = sprintf(
+							__( 'Your license key expired on %s.' ),
+							date_i18n( get_option( 'date_format' ), strtotime( $license_data->expires, current_time( 'timestamp' ) ) )
+						);
+						break;
+
+					case 'revoked' :
+
+						$message = __( 'Your license key has been disabled.' );
+						break;
+
+					case 'missing' :
+
+						$message = __( 'Invalid license.' );
+						break;
+
+					case 'invalid' :
+					case 'site_inactive' :
+
+						$message = __( 'Your license is not active for this URL.' );
+						break;
+
+					case 'item_name_mismatch' :
+
+						$message = sprintf( __( 'This appears to be an invalid license key for %s.' ), $args['name'] );
+						break;
+
+					case 'no_activations_left':
+
+						$message = __( 'Your license key has reached its activation limit.' );
+						break;
+
+					default :
+
+						$message = __( 'An error occurred, please try again.' );
+						break;
+				}
+
+				if ( ! empty( $message ) ) {
+					$base_url = admin_url( 'themes.php?page=' . $this->theme_slug . '-license' );
+					$redirect = add_query_arg( array( 'sl_theme_activation' => 'false', 'message' => urlencode( $message ) ), $base_url );
+
+					wp_redirect( $redirect );
+					exit();
+				}
+
+			}
+
+		}
 
 		// $response->license will be either "active" or "inactive"
 		if ( $license_data && isset( $license_data->license ) ) {
 			update_option( $this->theme_slug . '_license_key_status', $license_data->license );
 			delete_transient( $this->theme_slug . '_license_message' );
 		}
+
+		wp_redirect( admin_url( 'themes.php?page=' . $this->theme_slug . '-license' ) );
+		exit();
 
 	}
 
@@ -282,13 +355,40 @@ class EDD_Theme_Updater_Admin {
 			'item_name'  => urlencode( $this->item_name )
 		);
 
-		$license_data = $this->get_api_response( $api_params );
+		$response = $this->get_api_response( $api_params );
 
-		// $license_data->license will be either "deactivated" or "failed"
-		if ( $license_data && ( $license_data->license == 'deactivated' ) ) {
-			delete_option( $this->theme_slug . '_license_key_status' );
-			delete_transient( $this->theme_slug . '_license_message' );
+		// make sure the response came back okay
+		if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
+
+			if ( is_wp_error( $response ) ) {
+				$message = $response->get_error_message();
+			} else {
+				$message = __( 'An error occurred, please try again.' );
+			}
+
+		} else {
+
+			$license_data = json_decode( wp_remote_retrieve_body( $response ) );
+
+			// $license_data->license will be either "deactivated" or "failed"
+			if ( $license_data && ( $license_data->license == 'deactivated' ) ) {
+				delete_option( $this->theme_slug . '_license_key_status' );
+				delete_transient( $this->theme_slug . '_license_message' );
+			}
+
 		}
+
+		if ( ! empty( $message ) ) {
+			$base_url = admin_url( 'themes.php?page=' . $this->theme_slug . '-license' );
+			$redirect = add_query_arg( array( 'sl_theme_activation' => 'false', 'message' => urlencode( $message ) ), $base_url );
+
+			wp_redirect( $redirect );
+			exit();
+		}
+
+		wp_redirect( admin_url( 'themes.php?page=' . $this->theme_slug . '-license' ) );
+		exit();
+
 	}
 
 	/**
@@ -358,63 +458,83 @@ class EDD_Theme_Updater_Admin {
 			'url'        => home_url()
 		);
 
-		$license_data = $this->get_api_response( $api_params );
+		$response = $this->get_api_response( $api_params );
 
-		// If response doesn't include license data, return
-		if ( !isset( $license_data->license ) ) {
-			$message = $strings['license-unknown'];
-			return $message;
-		}
+		// make sure the response came back okay
+		if ( is_wp_error( $response ) || 200 !== wp_remote_retrieve_response_code( $response ) ) {
 
-		// We need to update the license status at the same time the message is updated
-		if ( $license_data && isset( $license_data->license ) ) {
-			update_option( $this->theme_slug . '_license_key_status', $license_data->license );
-		}
-
-		// Get expire date
-		$expires = false;
-		if ( isset( $license_data->expires ) ) {
-			$expires = date_i18n( get_option( 'date_format' ), strtotime( $license_data->expires ) );
-			$renew_link = '<a href="' . esc_url( $this->get_renewal_link() ) . '" target="_blank">' . $strings['renew'] . '</a>';
-		}
-
-		// Get site counts
-		$site_count = $license_data->site_count;
-		$license_limit = $license_data->license_limit;
-
-		// If unlimited
-		if ( 0 == $license_limit ) {
-			$license_limit = $strings['unlimited'];
-		}
-
-		if ( $license_data->license == 'valid' ) {
-			$message = $strings['license-key-is-active'] . ' ';
-			if ( $expires ) {
-				$message .= sprintf( $strings['expires%s'], $expires ) . ' ';
-			}
-			if ( $site_count && $license_limit ) {
-				$message .= sprintf( $strings['%1$s/%2$-sites'], $site_count, $license_limit );
-			}
-		} else if ( $license_data->license == 'expired' ) {
-			if ( $expires ) {
-				$message = sprintf( $strings['license-key-expired-%s'], $expires );
+			if ( is_wp_error( $response ) ) {
+				$message = $response->get_error_message();
 			} else {
-				$message = $strings['license-key-expired'];
+				$message = $strings['license-status-unknown'];
 			}
-			if ( $renew_link ) {
-				$message .= ' ' . $renew_link;
-			}
-		} else if ( $license_data->license == 'invalid' ) {
-			$message = $strings['license-keys-do-not-match'];
-		} else if ( $license_data->license == 'inactive' ) {
-			$message = $strings['license-is-inactive'];
-		} else if ( $license_data->license == 'disabled' ) {
-			$message = $strings['license-key-is-disabled'];
-		} else if ( $license_data->license == 'site_inactive' ) {
-			// Site is inactive
-			$message = $strings['site-is-inactive'];
+
 		} else {
-			$message = $strings['license-status-unknown'];
+
+			$license_data = json_decode( wp_remote_retrieve_body( $response ) );
+
+			// If response doesn't include license data, return
+			if ( !isset( $license_data->license ) ) {
+				$message = $strings['license-status-unknown'];
+				return $message;
+			}
+
+			// We need to update the license status at the same time the message is updated
+			if ( $license_data && isset( $license_data->license ) ) {
+				update_option( $this->theme_slug . '_license_key_status', $license_data->license );
+			}
+
+			// Get expire date
+			$expires = false;
+			if ( isset( $license_data->expires ) && 'lifetime' != $license_data->expires ) {
+				$expires = date_i18n( get_option( 'date_format' ), strtotime( $license_data->expires, current_time( 'timestamp' ) ) );
+				$renew_link = '<a href="' . esc_url( $this->get_renewal_link() ) . '" target="_blank">' . $strings['renew'] . '</a>';
+			} elseif ( isset( $license_data->expires ) && 'lifetime' == $license_data->expires ) {
+				$expires = 'lifetime';
+			}
+
+			// Get site counts
+			$site_count = $license_data->site_count;
+			$license_limit = $license_data->license_limit;
+
+			// If unlimited
+			if ( 0 == $license_limit ) {
+				$license_limit = $strings['unlimited'];
+			}
+
+			if ( $license_data->license == 'valid' ) {
+				$message = $strings['license-key-is-active'] . ' ';
+				if ( isset( $expires ) && 'lifetime' != $expires ) {
+					$message .= sprintf( $strings['expires%s'], $expires ) . ' ';
+				}
+				if ( isset( $expires ) && 'lifetime' == $expires ) {
+					$message .= $strings['expires-never'];
+				}
+				if ( $site_count && $license_limit ) {
+					$message .= sprintf( $strings['%1$s/%2$-sites'], $site_count, $license_limit );
+				}
+			} else if ( $license_data->license == 'expired' ) {
+				if ( $expires ) {
+					$message = sprintf( $strings['license-key-expired-%s'], $expires );
+				} else {
+					$message = $strings['license-key-expired'];
+				}
+				if ( $renew_link ) {
+					$message .= ' ' . $renew_link;
+				}
+			} else if ( $license_data->license == 'invalid' ) {
+				$message = $strings['license-keys-do-not-match'];
+			} else if ( $license_data->license == 'inactive' ) {
+				$message = $strings['license-is-inactive'];
+			} else if ( $license_data->license == 'disabled' ) {
+				$message = $strings['license-key-is-disabled'];
+			} else if ( $license_data->license == 'site_inactive' ) {
+				// Site is inactive
+				$message = $strings['site-is-inactive'];
+			} else {
+				$message = $strings['license-status-unknown'];
+			}
+
 		}
 
 		return $message;
@@ -448,3 +568,30 @@ class EDD_Theme_Updater_Admin {
 	}
 
 }
+
+/**
+ * This is a means of catching errors from the activation method above and displyaing it to the customer
+ */
+function edd_sample_theme_admin_notices() {
+	if ( isset( $_GET['sl_theme_activation'] ) && ! empty( $_GET['message'] ) ) {
+
+		switch( $_GET['sl_theme_activation'] ) {
+
+			case 'false':
+				$message = urldecode( $_GET['message'] );
+				?>
+				<div class="error">
+					<p><?php echo $message; ?></p>
+				</div>
+				<?php
+				break;
+
+			case 'true':
+			default:
+
+				break;
+
+		}
+	}
+}
+add_action( 'admin_notices', 'edd_sample_theme_admin_notices' );
